Updated BrazeEngine for compatibility with FactualEngine 9.6
