+++ demo/src/main/java/com/factual/engine/braze/helloworld/MainActivity.java
@@ -10,16 +10,10 @@ import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 
 import com.appboy.support.AppboyLogger;
-import com.factual.FactualCircumstance;
-import com.factual.FactualConfigMetadata;
-import com.factual.FactualError;
-import com.factual.FactualException;
-import com.factual.FactualInfo;
-import com.factual.engine.FactualClientListener;
+import com.factual.engine.api.FactualException;
 import com.factual.engine.FactualEngine;
-import com.factual.engine.braze.BrazeEngineIntegration;
 
-public class MainActivity extends AppCompatActivity implements FactualClientListener {
+public class MainActivity extends AppCompatActivity {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -42,7 +36,7 @@ public class MainActivity extends AppCompatActivity implements FactualClientList
     Log.i("engine", "starting initialization");
     Resources res = getResources();
     FactualEngine.initialize(getApplicationContext(), res.getString(R.string.com_factual_engine_api_key));
-    FactualEngine.setListener(this);
+    FactualEngine.setReceiver(ExampleFactualClientReceiver.class);
     Log.i("engine", "initialization complete");
   }
 
@@ -86,48 +80,4 @@ public class MainActivity extends AppCompatActivity implements FactualClientList
         0);
   }
 
-  /*********************************** FactualClientListener **************************************/
-
-  @Override
-  public void onStarted() {
-    Log.i("engine", "Engine has started.");
-    BrazeEngineIntegration.initializeBrazeEngineIntegration(getApplicationContext());
-  }
-
-  @Override
-  public void onStopped() {
-    Log.i("engine", "Engine has stopped.");
-  }
-
-  @Override
-  public void onError(FactualError e) {
-    Log.i("engine", e.getMessage());
-  }
-
-  @Override
-  public void onInfo(FactualInfo i) {
-    Log.i("engine", i.getInfo());
-  }
-
-  @Override
-  public void onSyncWithGarageComplete() {
-    Log.i("engine", "Garage synced.");
-  }
-
-  @Override
-  public void onConfigLoad(FactualConfigMetadata data) {
-    Log.i("engine", "Garage config loaded at: " + data.getVersion());
-    if (data.getGarageRelease() != null) {
-      for (FactualCircumstance circumstance : data.getGarageRelease().getCircumstances()){
-        Log.i("engine", "loaded circumstance: " + circumstance.getCircumstanceId());
-      }
-    } else {
-      Log.i("engine", "Garage release is empty");
-    }
-  }
-
-  @Override
-  public void onDiagnosticMessage(String diagnosticMessage) {
-    Log.i("engine", diagnosticMessage);
-  }
 }
