+++ library/build.gradle
@@ -30,9 +30,9 @@ repositories {
 
 ext {
     brazeVersion = "3.0.0"
-    engineVersion = "[8.4.0,)"
+    engineVersion = "[9.6.0,)"
     artifact = "braze-engine"
-    libraryVersion = "1.3.0"
+    libraryVersion = "1.4.0"
 }
 
 dependencies {
